+++ Landoria.CharacterVault/ProfileTransferService.cs
@@ -1,5 +1,3 @@
-extern alias ModSentryApi;
-
 using System;
 using System.Collections;
 using System.Collections.Generic;
@@ -7,8 +5,6 @@ using System.IO;
 using System.Linq;
 using System.Threading;
 using Landoria.SharedLib;
-using ModSentryGuestMarker =
-    ModSentryApi::Landoria.ModSentry.ModSentryGuestMarker;
 using UnityEngine;
 
 namespace Landoria.CharacterVault
@@ -124,13 +120,13 @@ namespace Landoria.CharacterVault
 
         internal bool ApproveGuest(ZRpc rpc)
         {
-            if (!ModSentryGuestMarker.IsMarked(rpc))
+            if (!CharacterGuestApi.IsGuest(rpc))
             {
                 return Approve(rpc);
             }
 
             CharacterVaultPlugin.Log.LogInfo(
-                "Recognized a ModSentry guest session; skipping character validation, " +
+                "Recognized a guest session; skipping character validation, " +
                 "vault session creation, profile import, and persistence.");
             return true;
         }
@@ -296,9 +292,9 @@ namespace Landoria.CharacterVault
 
         internal KickSaveEligibility GetKickSaveEligibility(ZNetPeer peer)
         {
-            if (peer?.m_rpc != null && ModSentryGuestMarker.IsMarked(peer.m_rpc))
+            if (peer?.m_rpc != null && CharacterGuestApi.IsGuest(peer.m_rpc))
             {
-                return KickSaveEligibility.ModSentryGuest;
+                return KickSaveEligibility.Guest;
             }
             if (peer?.m_rpc != null && !peer.IsReady())
             {